Add submit button to our text area
In an effort to deal with the problems with our callback happening at every key strong, we created a submit button with the intention that the callback would only happen when that was clicked instead of every time there was a key stroke.

This added a state to our component which was the current string.

It all seems to work (the right value is being returned when the button is clicked), *but* we're logging at every keystroke when I would have expected logging to only happen we click the submit button. Not sure why that is.

I also added a little to-do list that we need to look into.