Convert from getting videos to getting repositories
This starts the process of converting from the simple videos example in the Yew tutorial to actually getting repositories from the GitHub API. At the moment I'm just getting the id, name, and description of a repository, and sticking everything else in a `HashMap` using `serde(flatten)` (Thanks to @Wgaffa for sharing that idea.)

This doesn't work, yet, with something failing in the deserialization process. So I've started separating out the two steps (sending the request, and parsing the result) so we can more easily see what's happening, and so we can provide better feedback to users.