Switch to named repository components
This changes from using the `serde(flatten)` `HashMap` to using named components in the `Repository` struct. We had to use `Option<String>` for the `description` field because not all repositories have a non-null description.

We updated the HTML to only desplay the description if there is one, and to display a message saying there was no description if the description was null.