Add "Loading..." message
This adds a (very simple) "Loading..." message while we're waiting for our HTTP request to complete.

It's possible that the Yew Suspense tool would be a better choice here, but that would involve some non-trivial rewriting.

I also renamed "on_change" to "on_submit" at the subbestion of @DefineEvil on the live stream.