Add a `key` to the RepositoryList
As suggested by @DefineEvil, we added a `key` field to the `<RepositoryList ..>` tag in the top-level component. This forces the component to be destroyed and rebuilt when the organization changes, which appears to improve performance, and also ensures that the repository list is empty when the organization. This means we don't have to explicitly reset the repository list in `use_effect_with_deps`.