Turn on recommended Clippy warnings
This runs on the recommendations from the No Boilerplate video "Build your Rust lightsaber".

I couldn't figure out how to replace the if-let with Option::map_or_else as Clippy recommends, so that warning is still "live", but I fixed everything else except for the two `unwrap` warnings that definitely need to be resolved in the error handling improvements.