Work on pagination
This doesn't yet work, but it's a start on adding pagination of repositories.

We've had to expand the state of the `RepositoryList` component to include the current and last page numbers.

The last page number should be parseable from the `link` field in the response header, but that's not yet fully working. I tried using regex and got close but not quite there. @esitsu from Twitch suggested an alternative approach that parses the URLs using the url crate, and I might switch to that.

We restored the display of previously archived repositories because without them the pagination will be all confused. When we get this working, we probably do something to shuffle the archived repos to the end of the list.