Get the last page number parsing to work
The problem with the previous commit was that I was using `captures[0]`, which returned the _entire_ matched text and not just the capture group. Changing it to `captures[1]` fixed that.

I also changed all the `unwrap()`s to `expect()`s, but that's still not great and we really need to do something better there.