Replaced `match` in `repository_list()` with `map_or`
The `match` in `repository_list()` was kind of annoying because it was long, and had some non-trivial duplication. I realized that `map_or` could be used to clean that up. In doing so, I found that passing `parse_last_page` as an argument to `map_or` didn't work unless I changed its signature to take a `String` instead of a `&str`. I'm not thrilled about that change, but I'm not sure what to do about it. More research required.