Convert last page regex to static lazy expression
This converts the regular expression that is used to extract the last page from the GitHub response header to a static variable using `Lazy` so it only needs to be computed when it's actually first used.

We had to add the `once_cell` crate the dependencies, although it looks like this is being added to the language in `std` in an upcoming version.

I just used `unwrap()` on the construction of the Regex. Thanks to a suggestion from @Patatas_del_papa on Twitch, which suggested reading this post: https://blog.burntsushi.net/unwrap/