Only display pagination controls when there are at least 2 pages
This changes the display of the pagination controls so they only appear when there are at least two pages. It also fixes a bug where we were capturing `per_page=` when we wanted to just capture `page=`, which led to the `per_page` value being used as the last page number.