Add current page as paginator dependency
The `RepositoryPaginator` component needs to update whenever the `current_page` of the paginator state changes. This adds the current page as a dependency in `use_effect_with_deps()`.

I think this is correct or very nearly so, and that the problem we're having it due to parsing problems in my regex.