Comment understanding of `String` in properties
Add some comments documenting my (not awesome) understanding of the role of lifetimes in trying to replace `String` with `&str` in several of the components. Because we can't (currently) have generics in function components, so I don't think we can use `&str` in things like properties, which will bleed into components. Generics for function components will be included in the next version of Yew, though, so we might reconsider that if we upgrade to use that version of Yew down the road.