Convert from using regex to parse link string to string functions
The regex burned me twice by not working when I thought it did, so I decided to go back to esitsu@Twitch's suggestion to use string functions like `split()` and url parsing to manage all this.

This required bringing in the `url` crate, but meant we could drop the `once_cell` and `regex` crates.

This doesn't quite work because it panics if there's not 'rel="last"' entry in the link string, which happens when you request the last page. Our plan is to change this to return a `Result<Option<>, >` type to capture this possibility and the various errors that can occur.