Convert `parse_last_page` to return `Result<Option<T>, E>`
Instead of returning a `usize` directly (and basically hiding all the potential error issues, this changes `parse_past_page` to return `Result<Option<usize>, LinkParseError`.

The `Result` type is used here to indicate errors (of type `LinkParseError`), and the inner `Option` type is used to capture the possibility that we're on the last page, in which case there is not `rel="last"` entry.

There are three ways this can fail, which are captured in `LinkParseError`:

- `url::parse()` can fail (`UrlParseError`)
- There might not be a `rel="last"` (`PageEntryMissingError`)
- `parse::<usize>()` might fail (`PageNumParseError`)

We also implemented the `From` trait to convert from things like `url::ParseError` to the appropriate variant of `LinkParseError`.