Start to fix the call to `parse_last_page`
Changing `parse_last_page` to return `Result<Option<usize>, LinkParseError>` totally breaks the one place where we call it.

This is a total hack (we were deeply into guess and check at this point) to deal with that, and it's an ugly mess.

I think what we *really* want to do is actually capture the result, and do a `match` on it.

There's also an ownership issue in the definition of `parse_last_page` that needs to be addressed; I think that's probably "just" a clone().