Split the state into two structs
Having all the state info in one struct was potentially and helped lead to at least one error in the page button callback code.

This splits the state into two parts, one being the state for a page of repositories in the paginator, and the other being the state holding the map that records the desired archive status for repositories.