Passed the desired state archive up and down the chain
I needed to know the desired archive state down in the `RepositoryCard` component, so I ended up having to pass down a whole lot of information from the paginator to the list to the card, with more cloning along the way.

We tried as an experiment passing the `UseStateHandle` as a component property, but that didn't seem to have worked. At the very least we have a bug now that needs fixing that _may_ have been the cause of it?