Implement Yewdux state management
This creates two new Yewdux states: `Organization` and `ArchiveStateMap` which are read and set by the appropriate components.

We can't call hooks unless we're directly in a functional component, which means we can't use things like `use_store` in callbacks or spawn "threads", so we had to make sure to get the state tools in the components and pass clones of them into things like callbacks.

I'm hoping this significantly reduces the amount of cloning in the app, perhaps marginally improving performance.