Start `ReviewAndSubmit` component
This drafts a reasonable start on the `ReviewAndSubmit` component, which lists all the selected repositories.

At the moment it's missing any kind of header (oops) and doesn't have a "Submit" button that would trigger the actual archiving. It also violates the principle of least amazement in that if you unclick a repository it disappears and you have no way to get it back. :(