Refactor `bool` to `ArchiveState`
The `ArchiveStateMap` just used a simple `bool` to indicate whether the user wanted a repository to be archived or not. That didn't really work when we got to `ArchiveStateMap` because we needed to distinguish between the situation where a repo had been skipped in the pagination phase, or it had been unchecked in the "Review & Submit" phase.

So (and esitsu@Twitch gets a *ton* of the credit here) we created an `ArchiveState` enum with three variants, and some `from_` methods to convert booleans (the state of checkboxes) to appropriate `ArchiveState`s.

We then changed `ArchiveStateMap` to use the new `ArchiveState` type, and chased the (surprisingly small number of) compiler errors to fix any dependencies. And it worked!