Switch to "Prev"/"Next" UI
This switches from having separate buttons for each page of repositories to a "Prev"/"Next" approach that forces the user to go through all the pages before they can get to the "Review & Submit" page.

This also moves the buttons to the bottom of the page so that the user has to scroll through all the repositories before they can move on to the next page.