Add new `RepoId` type
This adds a type alias `RepoId` for `usize` to help us keep track of the various `usize` types we have floating around in this project.

I was suprised to find that this change didn't bleed out into the codebase hardly at all. I'd apparently managed to keep most of the references to repository IDs in this module, although that certainly hadn't been an intentional act on my part.

I also modified the `Props` for `RepositoryList` to take a `Vec<RepoId>` instead of a `Vec<Repository>`. If nothing else that should make the cloning there less expensive.