Add new `PageRepoMap` type
It turns out that we need to keep track of which pages we've loaded to avoid revisiting GitHub when we go back to a page or repositories that we've already looked at.

This new `PageRepoMap` is used to map from page numbers to vectors of repo IDs.

I also ended up adding a few more methods to `ArchiveStateMap` to get a repository by ID, and to get the _ID_s of repositories to review.