Change `RepositoryList` to take vec of IDs
This changes `RepositoryList` so it takes (in its properties) a `Vec<RepoId>` instead of `Vec<Repository>`. This works better with the new `PageRepoMap`, and also means we clone less when we clone a list of repositories since cloning `RepoId`s is presumably cheaper than cloning whole `Repository`s.

We also introduced an `Option` wrapper here to handle the possibility that a page of repositories hadn't yet loaded. I feel like that actually tidied up the logic here just a tad.