Remove vec of repos from paginator state
In order to avoid duplicate requests to GitHub for the same page, this removes the vector of repositories from the state of the paginator. Instead we use the `last_page` element of the state to get the appropriate vector of repository IDs. We did need to add `loaded` to the repository state, so when that changes the repository state will redraw. (Before that happened when the list of repositories updated.)