Add `AlreadyArchived` variant to `repository::DesiredState` enum
This adds a new `AlreadyArchived` variant to the `repository::DesiredState` (formerly `ArchiveState) enum. This was necessary to keep proper track of previously archived repositories.

In doing this, I changed a bunch of names in the hopes of making things more readable and better the names and the intent. This includes:

* ArchiveState => DesiredState
* Skip => Keep; SkippedInReview => KeptInReview
* ArchiveStateMap => StateMap

This currently compiles, but there is definitely the potential for issues as the VS Code renaming tool didn't always catch renamings when they appeared, for example, inside of macro calls.