Pass organization in to paginator as a property
Instead of having the organization as a piece of global state using Yewdux, this passes the organization into the paginator as a property. This is important because then we will make a new paginator when there is a new organization instead of having the same paginator persist from one organization to the next. Since the state of the paginator is entirely dependent on the organization, it makes sense to have all that reset when there's a new organization.