Make page map part of the internal state of the paginator
The page map is entirely local to the paginator and should not have been a "global" state using Yewdux. This makes that a lot state using the `state` hook.

There are also some renaming changes and we changed the `prev` and `next_or_review` callbacks so they send just the `has_seen_page` value instead of the whole page map to `make_button_callback`.