Remove dependency on `loaded` for `load_new_page`
The `use_effect_with_deps` used to depend on `loaded` (and `current_page`), but the dependency on `loaded` was unnecessary and (I think) causing problems with double downloading of pages from GitHub. This removes that dependency.

We also added an `if` test so that we only load the current page if we haven't already done so.