Add a dependency to the `page_map`
Having `load_new_page` called when the `page_map` changes fixes the problem of it not re-loading when we switched organizations.

The problem was that the two `use_effect_with_deps` calls are (presumably) happening asynchronously, so the one that depended on `current_page` was happening with the _old_ `page_map`. That was saying we'd already loaded that page, and we didn't load it on the new organization.

Adding the `page_map` as a dependency fixed that problem and I think all the loading is now correct!